requirements-local.txt
# Core Django & Database (Local Development)
Django==5.1.0
python-decouple==3.8

# GraphQL
graphene-django==3.2.0
django-graphql-jwt==0.4.0

# REST Framework & CORS
djangorestframework==3.15.2
django-cors-headers==4.4.0

# Background Tasks
celery==5.4.0
redis==5.0.1

# Media & Static Files
Pillow==10.4.0
whitenoise==6.7.0

# Development Tools
django-extensions==3.2.3

# Production Server
gunicorn==22.0.0

# Development Dependencies
black==24.8.0
isort==5.13.2
flake8==7.1.1
pytest==8.3.2
pytest-django==4.8.0
coverage==7.6.1
